+++ database/migrations/2021_05_09_155817_create_mine_rates_table.php
@@ -17,7 +17,7 @@ class CreateMineRatesTable extends Migration
             $table->id();
             $table->tinyInteger('mine');
             $table->tinyInteger('step');
-            $table->double('coeff',10,2);
+            $table->float('coeff',10,2);
             $table->timestamps();
         });
     }
